requirements.txt
charset-normalizer==2.1.1
idna==3.4
numpy==1.24.1
pandas==1.5.2
py-hanspell==1.1
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
six==1.16.0
urllib3==1.26.13
